reorganizing and adding javadoc
