+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/LocationResolverRegistry.java
