+++ projects/batfish/src/main/java/org/batfish/specifier/NodeNameRegexInterfaceLinkLocationSpecifierFactory.java
@@ -15,7 +15,7 @@ public class NodeNameRegexInterfaceLinkLocationSpecifierFactory
   }
 
   @Override
-  public LocationSpecifier specifierTyped(Pattern pattern) {
+  public LocationSpecifier buildLocationSpecifierTyped(Pattern pattern) {
     return new NodeNameRegexInterfaceLocationSpecifier(pattern);
   }
 }
