+++ projects/batfish/src/main/java/org/batfish/specifier/NodeNameRegexInterfaceLocationSpecifierFactory.java
@@ -10,7 +10,7 @@ public class NodeNameRegexInterfaceLocationSpecifierFactory
   }
 
   @Override
-  public LocationSpecifier specifierTyped(Pattern pattern) {
+  public LocationSpecifier buildLocationSpecifierTyped(Pattern pattern) {
     return new NodeNameRegexInterfaceLocationSpecifier(pattern);
   }
